+++ eio/tls_eio.mli
@@ -20,7 +20,11 @@ type t = [ `Tls | Eio.Flow.two_way_ty | Eio.Resource.close_ty ] r
 
     You must ensure a RNG is installed while using TLS, e.g. using [Mirage_crypto_rng_unix.use_default ()].
     Ideally, this would be part of the [server] config so you couldn't forget it,
-    but for now you'll get a runtime error if you forget. *)
+    but for now you'll get a runtime error if you forget.
+
+    Returns a TLS flow which may be fully active or half-closed.
+
+    @raise End_of_file if the peer closes before the handshake completes. *)
 val server_of_flow :
   Tls.Config.server ->
   [> Eio.Flow.two_way_ty | Eio.Resource.close_ty] r -> t
@@ -30,7 +34,11 @@ val server_of_flow :
 
     You must ensure a RNG is installed while using TLS, e.g. using [Mirage_crypto_rng_unix.use_default ()].
     Ideally, this would be part of the [client] config so you couldn't forget it,
-    but for now you'll get a runtime error if you forget. *)
+    but for now you'll get a runtime error if you forget.
+
+    Returns a TLS flow which may be fully active or half-closed.
+
+    @raise End_of_file if the peer closes before the handshake completes. *)
 val client_of_flow :
   Tls.Config.client -> ?host:[ `host ] Domain_name.t ->
   [> Eio.Flow.two_way_ty | Eio.Resource.close_ty] r -> t