+++ miou/tls_miou_unix.mli
@@ -77,20 +77,25 @@ val client_of_fd :
 (** [client_of_flow client ~host fd] is [t], after client-side TLS handshake of
     [fd] using [client] configuration and [host].
 
-    @raise End_of_file if we are not able to complete the handshake. *)
+    Returns a TLS flow which may be fully active or half-closed.
+
+    @raise End_of_file if the peer closes before the handshake completes. *)
 
 val server_of_fd :
   Tls.Config.server -> ?read_buffer_size:int -> Miou_unix.file_descr -> t
 (** [server_of_fd server fd] is [t], after server-side TLS handshake of [fd]
     using [server] configuration.
 
-    @raise End_of_file if we are not able to complete the handshake. *)
+    Returns a TLS flow which may be fully active or half-closed.
+
+    @raise End_of_file if the peer closes before the handshake completes. *)
 
 val connect : X509.Authenticator.t -> string * int -> t
 (** [connect authenticator (host, port)] is [t], a connected TLS connection
     to [host] on [port] using the default configuration and the
-    [authenticator]. *)
+    [authenticator].
+
+    Returns a TLS flow which may be fully active or half-closed. *)
 
 val epoch : t -> Tls.Core.epoch_data option
-(** [epoch t] returns [epoch], which contains information of the active
-    session. *)
+(** [epoch t] returns [epoch], which contains information of the session. *)